refactor: 프로젝트 패키지 구조 변경
