--- app/src/main/java/com/example/mz_focusnews/MyPageScreen.kt
@@ -1,285 +1,0 @@
-package com.example.mz_focusnews
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Surface
-import androidx.compose.material.Switch
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.example.mz_focusnews.ui.theme.Bg_Blue
-import com.example.mz_focusnews.ui.theme.Blue_400
-import com.example.mz_focusnews.ui.theme.Icon_Gray
-import com.example.mz_focusnews.ui.theme.preFontFamily
-
-@Composable
-fun MyPageScreen() {
-
-    var alarmChecked by remember { mutableStateOf(true) }
-    var locationChecked by remember { mutableStateOf(true) }
-
-
-    Surface(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Bg_Blue)
-            .padding(25.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Bg_Blue),
-            verticalArrangement = Arrangement.spacedBy(10.dp)
-        ) {
-
-            // UserInfoSection()
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 60.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                Image(
-                    painter = painterResource(R.drawable.example),
-                    contentDescription = "user image",
-                    modifier = Modifier
-                        .size(130.dp)
-                        .clip(RoundedCornerShape(100.dp))
-                        .background(Blue_400)
-                )
-                Text(
-                    modifier = Modifier.padding(top = 16.dp),
-                    text = "래로미",
-                    style = TextStyle(
-                        fontFamily = preFontFamily,
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 20.sp
-                    )
-                )
-                Text(
-                    modifier = Modifier.padding(top = 4.dp),
-                    text = "fofhal@gmail.com",
-                    style = TextStyle(
-                        fontFamily = preFontFamily,
-                        fontWeight = FontWeight.Medium,
-                        fontSize = 14.sp
-                    )
-                )
-            }
-
-            // QuizScoreSection()
-            Box(
-                modifier = Modifier.padding(top = 12.dp)
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(85.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(Color.White),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = "지금 내 상식 점수는?",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Medium,
-                            fontSize = 15.sp
-                        )
-                    )
-
-                    Text(
-                        modifier = Modifier.padding(top = 10.dp),
-                        text = "2,500 점",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 18.sp
-                        )
-                    )
-                }
-            }
-
-
-            // KeywordSettingSection()
-            Box(
-                modifier = Modifier.padding(vertical = 4.dp)
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(110.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(Color.White),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Text(
-                        text = "이런 뉴스가 좋아요!",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Medium,
-                            fontSize = 15.sp
-                        )
-                    )
-
-                    Text(
-                        modifier = Modifier.padding(top = 10.dp),
-                        text = "크리스피 도넛",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Black,
-                            fontSize = 20.sp
-                        )
-                    )
-                }
-            }
-
-            // RecentNewsList()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(52.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(Color.White),
-                contentAlignment = Alignment.CenterStart
-            ) {
-                Text(
-                    modifier = Modifier.padding(start = 25.dp),
-                    text = "최근 본 뉴스",
-                    style = TextStyle(
-                        fontFamily = preFontFamily,
-                        fontWeight = FontWeight.Medium,
-                        fontSize = 15.sp
-                    )
-                )
-            }
-
-            // FavoriteNewsList()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(52.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(Color.White),
-                contentAlignment = Alignment.CenterStart
-            ) {
-                Text(
-                    modifier = Modifier.padding(start = 25.dp),
-                    text = "내가 좋아하는 뉴스",
-                    style = TextStyle(
-                        fontFamily = preFontFamily,
-                        fontWeight = FontWeight.Medium,
-                        fontSize = 15.sp
-                    )
-                )
-            }
-
-            // AlarmToggleButton()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(52.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(Color.White),
-                contentAlignment = Alignment.CenterStart
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        modifier = Modifier.padding(start = 25.dp),
-                        text = "속보 알림 수신 동의",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Medium,
-                            fontSize = 15.sp
-                        )
-                    )
-
-                    Switch(
-                        modifier = Modifier.padding(start = 8.dp, bottom = 2.dp),
-                        checked = alarmChecked,
-                        onCheckedChange = { alarmChecked = it },
-                    )
-                }
-            }
-
-            // LocationToggleButton()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(52.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(Color.White),
-                contentAlignment = Alignment.CenterStart
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        modifier = Modifier.padding(start = 25.dp),
-                        text = "위치 정보 이용 동의",
-                        style = TextStyle(
-                            fontFamily = preFontFamily,
-                            fontWeight = FontWeight.Medium,
-                            fontSize = 15.sp
-                        )
-                    )
-
-                    Switch(
-                        modifier = Modifier.padding(start = 8.dp, bottom = 2.dp),
-                        checked = locationChecked,
-                        onCheckedChange = { locationChecked = it },
-                    )
-                }
-            }
-
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    text = "로그아웃",
-                    style = TextStyle(
-                        fontFamily = preFontFamily,
-                        fontWeight = FontWeight.Medium,
-                        fontSize = 15.sp,
-                        color = Icon_Gray
-                    )
-                )
-            }
-        }
-    }
-}
-
-@Preview
-@Composable
-fun MyPagePreview() {
-    MyPageScreen()
-}
