--- app/src/main/java/com/example/mz_focusnews/NewsApp.java
@@ -1,16 +1,0 @@
-package com.example.mz_focusnews;
-
-import android.app.Application;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * 뉴스 애플리케이션 클래스
- */
-public class NewsApp extends Application {
-    @Override
-    public void onCreate() {
-        super.onCreate();
-    }
-}
