+++ app/src/main/java/com/example/mz_focusnews/feature/favorite/FavorNewsScreen.kt
@@ -1,0 +1,27 @@
+package com.example.mz_focusnews.feature.favorite
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.mz_focusnews.core.theme.Bg_Blue
+
+@Composable
+fun FavorNewsScreen() {
+
+    Surface(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Bg_Blue)
+            .padding(25.dp)
+    ) {
+
+
+
+    }
+
+
+}
