+++ app/src/main/java/com/example/mz_focusnews/feature/home/BreakingSection.kt
@@ -1,0 +1,67 @@
+package com.example.mz_focusnews.feature.home
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.mz_focusnews.R
+import com.example.mz_focusnews.core.theme.Bg_Blue
+import com.example.mz_focusnews.core.theme.preFontFamily
+
+@Composable
+fun BreakingSection() {
+    Surface(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(top = 2.dp)
+    ) {
+        Box(
+            modifier = Modifier.fillMaxWidth().background(Bg_Blue)
+        ) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .background(Color.White, RoundedCornerShape(12.dp))
+                    .padding(horizontal = 14.dp, vertical = 13.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Image(
+                    painter = painterResource(R.drawable.breaking),
+                    contentDescription = "breaking news icon",
+                    modifier = Modifier.size(25.dp)
+                )
+                Text(
+                    modifier = Modifier.padding(start = 12.dp),
+                    text = "산림청 \"경북 영덕 산불 주불 진화\" 평균 진화율 94% 넘어",
+                    style = TextStyle(
+                        fontFamily = preFontFamily,
+                        fontWeight = FontWeight.SemiBold,
+                        fontSize = 12.5.sp
+                    )
+                )
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun BreakingPreview() {
+    BreakingSection()
+}
