--- app/src/main/java/com/example/mz_focusnews/ui/theme/color.kt
@@ -1,21 +1,0 @@
-package com.example.mz_focusnews.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Blue_100 = Color(0xFFFFFFFF) // #FFFFFF
-val Bg_Blue = Color(0xFFF9FBFF) // #F9FBFF
-val Blue_300 = Color(0xFFEAF4FF) // #EAF4FF
-val Blue_400 = Color(0xFFDAEDFE) // #DAEDFE
-val Category_Blue = Color(0xFFDEEEFF)
-val Blue_500 = Color(0xFFC4E3FF) // #C4E3FF
-val Today_Blue = Color(0xFFACD7FF) // #ACD7FF
-val Blue_700 = Color(0xFF89BAF5) // #89BAF5
-val Blue_800 = Color(0xFF63A4F2) // #63A4F2
-val Blue_900 = Color(0xFF3D8FEF) // #3D8FEF
-
-val Gray_100 = Color(0xFFE9EBED)
-val Gray_200 = Color(0xFFDBDBDB)
-val Gray_300 = Color(0xFFD1D1D1)
-val Gray_400 = Color(0xFFADADAD)
-val Gray_500 = Color(0xFF484848)
-val Icon_Gray = Color(0xFF8C8C8C)
